pyproject.toml
[tool.poetry]
name = "shop-api-rest"
version = "0.1.0"
description = ""
authors = ["alexhdz1 <alex.hdz.c@ciencias.unam.mx>"]

[tool.poetry.dependencies]
python = "^3.9"
asgiref = "3.3.1"
certifi = "2020.12.5"
chardet = "4.0.0"
coreapi = "2.3.3"
coreschema = "0.0.4"
Django = "3.1.5"
django-cors-headers = "3.7.0"
djangorestframework = "3.12.2"
djangorestframework-simplejwt = "4.6.0"
drf-yasg = "1.20.0"
idna = "2.10"
inflection = "0.5.1"
itypes = "1.2.0"
Jinja2 = "2.11.3"
MarkupSafe = "1.1.1"
openapi-codec = "1.3.2"
packaging = "20.9"
Pillow = "8.1.0"
PyJWT = "2.0.1"
pyparsing = "2.4.7"
pytz = "2020.5"
requests = "2.25.1"
"ruamel.yaml" = "0.16.12"
simplejson = "3.17.2"
sqlparse = "0.4.1"
uritemplate = "3.0.1"
urllib3 = "1.26.3"
gevent = "^22.10.2"
django-environ = "^0.10.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
